Add more integration tests and fix some bugs. Add client secret validation to only accept guid strings.

+++ src/api/Fig.Api/Services/SettingsService.cs
@@ -1,6 +1,7 @@
 using Fig.Api.Converters;
 using Fig.Api.Datalayer.BusinessEntities;
 using Fig.Api.Datalayer.Repositories;
+using Fig.Api.Exceptions;
 using Fig.Api.ExtensionMethods;
 using Fig.Contracts.SettingDefinitions;
 using Fig.Contracts.Settings;
@@ -9,13 +10,13 @@ namespace Fig.Api.Services;
 
 public class SettingsService : ISettingsService
 {
+    private readonly IEventLogFactory _eventLogFactory;
+    private readonly IEventLogRepository _eventLogRepository;
     private readonly ILogger<SettingsService> _logger;
     private readonly ISettingClientRepository _settingClientRepository;
-    private readonly IEventLogRepository _eventLogRepository;
-    private readonly ISettingHistoryRepository _settingHistoryRepository;
     private readonly ISettingConverter _settingConverter;
     private readonly ISettingDefinitionConverter _settingDefinitionConverter;
-    private readonly IEventLogFactory _eventLogFactory;
+    private readonly ISettingHistoryRepository _settingHistoryRepository;
 
     public SettingsService(ILogger<SettingsService> logger,
         ISettingClientRepository settingClientRepository,
@@ -39,24 +40,25 @@ public class SettingsService : ISettingsService
         var existingRegistrations = _settingClientRepository.GetAllInstancesOfClient(client.Name).ToList();
 
         if (IsAlreadyRegisteredWithDifferentSecret())
-        {
             throw new UnauthorizedAccessException(
                 "Settings for that service have already been registered with a different secret.");
-        }
 
         var clientBusinessEntity = _settingDefinitionConverter.Convert(client);
 
+        if (clientBusinessEntity.Settings.Any(a => !a.Isvalid())) throw new InvalidSettingException();
+
         clientBusinessEntity.ClientSecret = clientSecret;
-        
+
         if (!existingRegistrations.Any())
         {
             _settingClientRepository.RegisterClient(clientBusinessEntity);
             RecordSettingValues(clientBusinessEntity);
-            _eventLogRepository.Add(_eventLogFactory.InitialRegistration(clientBusinessEntity.Id, clientBusinessEntity.Name));
+            _eventLogRepository.Add(
+                _eventLogFactory.InitialRegistration(clientBusinessEntity.Id, clientBusinessEntity.Name));
         }
         else if (existingRegistrations.All(x => x.HasEquivalentDefinitionTo(clientBusinessEntity)))
         {
-            foreach(var registration in existingRegistrations)
+            foreach (var registration in existingRegistrations)
                 _eventLogRepository.Add(_eventLogFactory.IdenticalRegistration(registration.Id, registration.Name));
         }
         else
@@ -65,7 +67,8 @@ public class SettingsService : ISettingsService
             foreach (var updatedDefinition in existingRegistrations)
             {
                 _settingClientRepository.UpdateClient(updatedDefinition);
-                _eventLogRepository.Add(_eventLogFactory.UpdatedRegistration(updatedDefinition.Id, updatedDefinition.Name));
+                _eventLogRepository.Add(
+                    _eventLogFactory.UpdatedRegistration(updatedDefinition.Id, updatedDefinition.Name));
             }
         }
 
@@ -79,10 +82,7 @@ public class SettingsService : ISettingsService
     public IEnumerable<SettingsClientDefinitionDataContract> GetAllClients()
     {
         var settings = _settingClientRepository.GetAllClients();
-        foreach (var setting in settings)
-        {
-            yield return _settingDefinitionConverter.Convert(setting);
-        }
+        foreach (var setting in settings) yield return _settingDefinitionConverter.Convert(setting);
     }
 
     public IEnumerable<SettingDataContract> GetSettings(string clientName, string clientSecret, string? instance)
@@ -90,81 +90,72 @@ public class SettingsService : ISettingsService
         var existingRegistration = _settingClientRepository.GetClient(clientName, instance);
 
         if (existingRegistration != null && existingRegistration.ClientSecret != clientSecret)
-        {
             throw new UnauthorizedAccessException();
-        }
 
-        if (existingRegistration == null)
-        {
-            throw new KeyNotFoundException();
-        }
+        if (existingRegistration == null) throw new KeyNotFoundException();
 
-        foreach (var setting in existingRegistration.Settings)
-        {
-            yield return _settingConverter.Convert(setting);
-        }
+        foreach (var setting in existingRegistration.Settings) yield return _settingConverter.Convert(setting);
     }
 
     public void DeleteClient(string clientName, string? instance)
     {
         var client = _settingClientRepository.GetClient(clientName, instance);
-        if (client != null)
-        {
-            _settingClientRepository.DeleteClient(client);
-        }
+        if (client != null) _settingClientRepository.DeleteClient(client);
     }
 
-    public void UpdateSettingValues(string id, string? instance,
+    public void UpdateSettingValues(string clientName, string? instance,
         IEnumerable<SettingDataContract> updatedSettings)
     {
-        bool dirty = false;
-        var client = _settingClientRepository.GetClient(id, instance);
+        var dirty = false;
+        var client = _settingClientRepository.GetClient(clientName, instance);
 
         if (client == null)
         {
-            var nonOverrideClient = _settingClientRepository.GetClient(id);
+            var nonOverrideClient = _settingClientRepository.GetClient(clientName);
 
-            if (nonOverrideClient == null)
-            {
-                return;
-            }
+            if (nonOverrideClient == null) throw new InvalidClientException(clientName);
 
             client = nonOverrideClient.CreateOverride(instance);
+            _settingClientRepository.RegisterClient(client);
             dirty = true;
         }
 
         var updatedSettingBusinessEntities = updatedSettings.Select(a => _settingConverter.Convert(a));
-        
+        var eventLogs = new List<EventLogBusinessEntity>();
         foreach (var updatedSetting in updatedSettingBusinessEntities)
         {
             var setting = client.Settings.FirstOrDefault(a => a.Name == updatedSetting.Name);
-            
+
             if (setting != null && updatedSetting.ValueAsJson != setting.ValueAsJson)
             {
                 var originalValue = setting.Value;
                 setting.Value = updatedSetting.Value;
-                RegisterSettingValueChanged(client.Id, 
+                eventLogs.Add(_eventLogFactory.SettingValueUpdate(client.Id,
                     client.Name,
                     instance,
                     setting.Name,
                     originalValue,
-                    updatedSetting.Value);
+                    updatedSetting.Value));
                 dirty = true;
             }
         }
 
+        if (client.Settings.Any(a => !a.Isvalid())) throw new InvalidSettingException();
+
         if (dirty)
         {
             _settingClientRepository.UpdateClient(client);
+            foreach (var eventLog in eventLogs)
+                _eventLogRepository.Add(eventLog);
         }
     }
-    
+
     private void UpdateRegistrationsWithNewDefinitions(
-        SettingClientBusinessEntity updatedSettingDefinitions, 
+        SettingClientBusinessEntity updatedSettingDefinitions,
         List<SettingClientBusinessEntity> existingRegistrations)
     {
         var settingValues = existingRegistrations.ToDictionary(
-            a => a.Instance ?? "Default", 
+            a => a.Instance ?? "Default",
             b => b.Settings.ToList());
 
         foreach (var registration in existingRegistrations)
@@ -175,10 +166,7 @@ public class SettingsService : ISettingsService
             {
                 var newSetting = setting.Clone();
                 var matchingSetting = values.FirstOrDefault(a => a.Name == newSetting.Name);
-                if (matchingSetting != null)
-                {
-                    newSetting.Value = matchingSetting.Value;
-                }
+                if (matchingSetting != null) newSetting.Value = matchingSetting.Value;
                 registration.Settings.Add(newSetting);
             }
         }
@@ -187,8 +175,7 @@ public class SettingsService : ISettingsService
     private void RecordSettingValues(SettingClientBusinessEntity client)
     {
         foreach (var setting in client.Settings)
-        {
-            _settingHistoryRepository.Add(new SettingValueBusinessEntity()
+            _settingHistoryRepository.Add(new SettingValueBusinessEntity
             {
                 // TODO check if this is populated here
                 ClientId = client.Id,
@@ -196,21 +183,5 @@ public class SettingsService : ISettingsService
                 SettingName = setting.Name,
                 Value = setting.Value
             });
-        }
-    }
-
-    private void RegisterSettingValueChanged(Guid clientId,
-        string clientName,
-        string? instance,
-        string settingName,
-        object originalValue,
-        object newValue)
-    {
-        _eventLogRepository.Add(_eventLogFactory.SettingValueUpdate(clientId,
-            clientName,
-            instance,
-            settingName,
-            originalValue,
-            newValue));
     }
 }
